Internships/requirements.txt
Flask
python-docx
PyPDF2==2.12.1
textract
requests
apscheduler